components/backend/scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["John Shanahan <shanahan.jrs@gmail.com>"]

[tool.poetry.dependencies]
analysis_utils = { path = "../../modules/analysis_utils/" }
python = "^3.6"
python-twitter = "^3.5"
flask = "^1.1.2"
redis = "^3.5.3"
praw = "^7.1.0"
newsapi-python = "^0.2.6"

[tool.poetry.scripts]
serve = 'scraper.app:run'

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
